% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co.R
\docType{data}
\name{co}
\alias{co}
\title{Tibble con i dati di co per 14 stazioni della regione LIGURIA}
\format{
Un tibble con 8 colonne e 10234 osservazioni
}
\usage{
co
}
\description{
Tibble con i dati di co per 14 stazioni della regione LIGURIA
}
\keyword{datasets}
